// src/interfaces/IHomeTranslate.ts
export interface IHomeTranslate {
	setupTrip: string;
	oneWay: string;
	roundTrip: string;
	startPlace: string;
	placeholderStart: string;
	endPlace: string;
	placeholderEnd: string;
	startDate: string;
	amountTickets: string;
	findTrip: string;
	howToBuyTicket: string;
	rules: string;
	contact: string;
	headquarters: string;
	callNow: string;
	service: string;
	serviceUs: string;
	news: string;
	seeAll: string;
	seeMore: string;
	placeholderSearchCity: string;
	startPosition: string;
	searchRecent: string;
	seen: string;
	comment: string;
	tourGoodPrice: string;
	onlyWith: string;
	guestBooked: string;
	cityDistrict: string;
	locationNotFound: string;
	chooseTicket: string;
	pickStartDate: string;
	readPost: string;
	endDate: string;
	pickEndDate: string;
	pickingDay: string;
}